+++ src/test/java/lotto/domain/ticket/LottoTicketTest.java
@@ -1,6 +1,6 @@
 package lotto.domain.ticket;
 
-import lotto.domain.ticket.ball.LottoNumber;
+import lotto.domain.ticket.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
